Added backend complete for operation type. #66 #67 #68 #69

+++ src/G74/Domain/Aggregates/OperationRequest/OperationRequest.cs
@@ -8,11 +8,11 @@ public class OperationRequest : IAggregateRoot
     public long OperationRequestId {get; set; }
     public MedicalRecordNumber MedicalRecordNumber {get; set; }
     public LicenceNumber LicenceNumber {get; set; }
-    public long OperationTypeId {get; set; }
+    public int OperationTypeId {get; set; }
     public DeadlineDate DeadlineDate {get; set; }
     public Priority Priority {get; set; }
 
-    public OperationRequest(MedicalRecordNumber medicalRecordNumber, LicenceNumber licenceNumber, long operationTypeId, DeadlineDate deadlineDate, Priority priority)
+    public OperationRequest(MedicalRecordNumber medicalRecordNumber, LicenceNumber licenceNumber, int operationTypeId, DeadlineDate deadlineDate, Priority priority)
     {
         MedicalRecordNumber = medicalRecordNumber;
         LicenceNumber = licenceNumber;
@@ -20,7 +20,7 @@ public class OperationRequest : IAggregateRoot
         DeadlineDate = deadlineDate;
         Priority = priority;
     }
-    public OperationRequest(MedicalRecordNumber medicalRecordNumber, LicenceNumber licenceNumber, long operationTypeId, DeadlineDate deadlineDate, Priority priority, long operationRequestId)
+    public OperationRequest(MedicalRecordNumber medicalRecordNumber, LicenceNumber licenceNumber, int operationTypeId, DeadlineDate deadlineDate, Priority priority, long operationRequestId)
     {
         MedicalRecordNumber = medicalRecordNumber;
         LicenceNumber = licenceNumber;
